Moved the BasicEvent tests to their own source file
